modelling_3d/fretHand/MANO/requirements.txt
chumpy==0.70
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
glfw==2.7.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.0
Panda3D==1.10.14
panda3d-gltf==1.2.0
panda3d-simplepbr==0.12.0
pillow==10.3.0
pyglet==2.0.15
PyOpenGL==3.1.7
pyparsing==3.1.2
pyperclip==1.8.2
python-dateutil==2.9.0.post0
scipy==1.13.0
screeninfo==0.8.1
setuptools==69.5.1
six==1.16.0
sympy==1.12
torch==2.2.2
trimesh==4.3.1
typing_extensions==4.11.0
ursina==6.1.2
wheel==0.43.0
PyYAML